4.2 to 5.0 merge

+++ src/Raven.Server/Documents/Expiration/ExpiredDocumentsCleaner.cs
@@ -199,7 +199,7 @@ namespace Raven.Server.Documents.Expiration
                 _forExpiration = forExpiration;
             }
 
-            protected override int ExecuteCmd(DocumentsOperationContext context)
+            protected override long ExecuteCmd(DocumentsOperationContext context)
             {
                 DeletionCount =
                     _forExpiration
