+++ src/Raven.Server/Documents/Handlers/BatchHandler.cs
@@ -501,7 +501,7 @@ namespace Raven.Server.Documents.Handlers
                 _batch = batch;
             }
 
-            protected override int ExecuteCmd(DocumentsOperationContext context)
+            protected override long ExecuteCmd(DocumentsOperationContext context)
             {
                 var global = context.LastDatabaseChangeVector ??
                              (context.LastDatabaseChangeVector = DocumentsStorage.GetDatabaseChangeVector(context));
@@ -734,7 +734,7 @@ namespace Raven.Server.Documents.Handlers
                 return false;
             }
 
-            protected override int ExecuteCmd(DocumentsOperationContext context)
+            protected override long ExecuteCmd(DocumentsOperationContext context)
             {
                 if (IsClusterTransaction)
                 {
