+++ src/Raven.Server/Documents/Replication/IncomingReplicationHandler.cs
@@ -897,7 +897,7 @@ namespace Raven.Server.Documents.Replication
                 _trigger = trigger;
             }
 
-            protected override int ExecuteCmd(DocumentsOperationContext context)
+            protected override long ExecuteCmd(DocumentsOperationContext context)
             {
                 var operationsCount = 0;
                 var lastReplicatedEtag = DocumentsStorage.GetLastReplicatedEtagFrom(context, _sourceDatabaseId);
@@ -951,7 +951,7 @@ namespace Raven.Server.Documents.Replication
                 _lastEtag = lastEtag;
             }
 
-            protected override int ExecuteCmd(DocumentsOperationContext context)
+            protected override long ExecuteCmd(DocumentsOperationContext context)
             {
                 try
                 {
